Add files via upload
A task list is a prioritized set of activities you (or your team) need to do to complete a project.

A task list contains specific tasks assigned to a task and when the task needs to be done. Typically these lists are created at the start of a project. And are agreed upon by the entire team and the stakeholders.

For a small project, a task list can be a simple checklist. However, for work-related projects, a task list is often mapped to a project and resource timeline. This ensures the timely availability of resources to execute a task and deliver the project on time.